prompt-manager/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.3
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
pyyaml==6.0.2
alembic==1.14.1
